wip: push Initialize into initializestep
